chore: fix `flake.nix` workable (#1571)
* chore: format flake.nix by nixfmt

* chore: ignore auto-generated files to develop with nix flake

* chore: common settings that are useful for using nix flake

* chore: add cache server for GHC

* chore: bump `flake.lock`

* chore: replace flake.nix by haskell.nix

It is by far simpler to write using haskell.nix.
It is because it automatically draws an analogy from cabal files and so on.
There is no need to be afraid of increasing dependencies if using `flake.nix`,
even if it is `default.nix` which is referenced by other nix projects.
And upgrade nix GHC version 9.6.6.

* chore: update `flake.lock` with new haskell.nix

* chore: add `projectFileName` to flake.nix

project type is cabal.

* chore: add `postgresql` to `buildInputs`

sqlite is not required for the build, but postgresql was.

* chore: add pkg-config flag for postgresql-libpq

* chore: fix mysql depend

Now `direnv allow` is successfully completed.

* fix: use https protocol instead of git protocol

Fix the following warnings.

```
Warning: [git-protocol] Cloning over git:// might lead to an arbitrary code
execution vulnerability. Furthermore, popular forges like GitHub do not
support it. Use https:// or ssh:// instead.
```

* chore: fix indent

```
Warning: persistent-mongoDB.cabal:44:4: Inconsistent indentation. Indentation
jumps at lines 44
```

* chore: fix indent

* chore: add depend for test

Since MongoDB is SSPL,
I have not added it for now because of the UNFREE warning.

* chore: ignore `.envrc` by `.gitignore`

People who use direnv for personal purposes and people who use nix-direnv should be able to coexist once gitignore is done.
Those who want to set up their own personal setup should know what commands they should write.
It is impossible to provide a setup that satisfies everyone, since every individual's setup needs should be different.
Therefore, in this case, gitignore and leave it to the individual to decide.